+++ packages/remix-entry-server/package.json
@@ -2,20 +2,38 @@
   "name": "@vercel/remix-entry-server",
   "version": "0.0.1",
   "description": "Isomorphic `entry.server` implementation for Vercel's Serverless and Edge runtimes",
-  "main": "./entry.server.node.mjs",
+  "homepage": "https://vercel.com/docs",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/vercel/vercel.git",
+    "directory": "packages/remix-entry-server"
+  },
+  "main": "./dist/entry.server.node.js",
   "exports": {
     ".": {
-      "node": "./entry.server.node.mjs",
-      "edge-light": "./entry.server.edge.mjs",
-      "worker": "./entry.server.edge.mjs",
-      "browser": "./entry.server.edge.mjs"
+      "node": "./dist/entry.server.node.js",
+      "edge-light": "./dist/entry.server.edge.js",
+      "worker": "./dist/entry.server.edge.js",
+      "browser": "./dist/entry.server.edge.js"
     },
     "./package.json": "./package.json"
   },
+  "scripts": {
+    "build": "node build.js",
+    "test-integration-once": "pnpm test test/integration.test.ts",
+    "test": "jest --env node --verbose --bail --runInBand"
+  },
   "license": "MIT",
+  "files": [
+    "dist"
+  ],
   "dependencies": {
     "@remix-run/node": "1.12.0",
     "isbot": "3.6.5",
     "react-dom": "18.2.0"
+  },
+  "devDependencies": {
+    "@types/react-dom": "18.0.10",
+    "typescript": "4.9.4"
   }
 }
