+++ packages/remix-entry-server/src/entry.server.edge.ts
@@ -1,0 +1,27 @@
+import isbot from 'isbot';
+import { renderToReadableStream } from 'react-dom/server';
+
+export default async function handleRequest(
+  request: Request,
+  responseStatusCode: number,
+  responseHeaders: Headers,
+  remixServer: any
+) {
+  const body = await renderToReadableStream(remixServer, {
+    signal: request.signal,
+    onError(error) {
+      console.error(error);
+      responseStatusCode = 500;
+    },
+  });
+
+  if (isbot(request.headers.get('user-agent'))) {
+    await body.allReady;
+  }
+
+  responseHeaders.set('Content-Type', 'text/html');
+  return new Response(body, {
+    headers: responseHeaders,
+    status: responseStatusCode,
+  });
+}
